Refactor GameLevel Menu : - SQL : Moves ID from BIGINT to TINYINT, for menuid and levelid primary keys - Model : Moves ID from Long to Byte, for menuid and levelid primary keys

+++ src/main/java/fr/softsf/sudokufx/dto/GameLevelDto.java
@@ -6,7 +6,7 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 
 public record GameLevelDto(
-        Long levelid,
+        Byte levelid,
         @NotNull
         @Min(1)
         @Max(3)
